Sleep for 14ms on Linux while video is paused
Since vsync is unreliable on linux, we let mpv wait on video frames instead. However, mpv does not wait if the video is paused, so when the video is paused we spin at 100% CPU again. To fix this, just wait 14ms. This leaves 2ms for moov to update, which is more than enough. By hardcoding 14ms, i.e. an update time target of 16ms, we assume a 60hz refresh rate, so if the display has a higher refresh rate, the UI will be rendered sub-optimally, but it's better than nothing.